Versión correcta algoritmo cuadrantes

+++ app/Http/Controllers/CuadrantesController.php
@@ -194,7 +194,11 @@ class CuadrantesController extends Controller
 
 	    //Eliminamos todos aquellos servicios almacenados en la base de datos cuya fecha sea mayor a hoy
 	    $eliminarAPartirDe = new Carbon();
-		$eliminarAPartirDe->setDate(2016, 11, 20);
+	    //$eliminarAPartirDe->setDate(2016, 11, 20);
+	    if ($eliminarAPartirDe->dayOfWeek == Carbon::SATURDAY || $eliminarAPartirDe->dayOfWeek == Carbon::SUNDAY) {
+		    echo "El algoritmo no puede lanzarse ni un sábado ni un domingo";
+		    return;
+	    }
 		$this->cuadranteRepository->deleteAllAfterDate($eliminarAPartirDe);
 
 	    for ($i=0; $i<8;$i++) {
@@ -214,7 +218,7 @@ class CuadrantesController extends Controller
 
 				    $now = new Carbon();
 				    $now->setTime(0, 0, 0);
-					$now->setDate(2016, 11, 21);
+					//$now->setDate(2016, 11, 21);
 				    $now->addWeeks( $i );
 				    if ($period == 4) continue;
 				    foreach ( $weekdays[ $period ] as $weekday ) {
@@ -222,7 +226,7 @@ class CuadrantesController extends Controller
 					    echo "Miramos el día {$weekday->value} {$now->day}<br>";
 					    if (!$now->isFuture()) {
 					    	echo "No analizamos el día {$now->day} porque es pasado<br>";
-						    //continue;
+						    continue;
 					    }
 
 					    if ( !isset( $servicesConditions[ $period ][ $group ] ) ) {
